+++ packages/language-js/src/expressions/query-selectors.ts
@@ -94,7 +94,7 @@ function newSelectorAsStandaloneStringExpressions():
  * @param options The {@link QuerySelectorOptions}.
  * @returns A set of {@link MangleExpression}s.
  */
-function querySelectorExpressionFactory(
+function fallback(
   options: QuerySelectorOptions,
 ): Iterable<MangleExpression> {
   const config: QuerySelectorConfig = {
@@ -110,6 +110,101 @@ function querySelectorExpressionFactory(
   }
 
   return result;
+}
+
+/**
+ * Get {@link MangleExpression}s to match query selectors for attributes in
+ * JavaScript, e.g. `foobar` in `[foobar] { }`.
+ *
+ * @returns The {@link MangleExpression}s.
+ */
+function newAttributeSelectorExpression(): Iterable<MangleExpression> {
+  return [
+    ...newQuerySelectorExpressions({
+      kind: "attribute",
+      prefix: /\[\s*/.source,
+      suffix: /\s*(?:]|=|~=|\|=|\^=|\$=|\*=)/.source,
+    }),
+    ...newSelectorAsStandaloneStringExpressions(),
+  ];
+}
+
+/**
+ * Get {@link MangleExpression}s to match query selectors for classes in
+ * JavaScript, e.g. `foobar` in `.foobar { }`.
+ *
+ * @returns The {@link MangleExpression}s.
+ */
+function newClassSelectorExpression(): Iterable<MangleExpression> {
+  return [
+    ...newQuerySelectorExpressions({
+      kind: "class",
+      prefix: /\./.source,
+    }),
+    ...newSelectorAsStandaloneStringExpressions(),
+  ];
+}
+
+/**
+ * Get {@link MangleExpression}s to match query selectors for elements in
+ * JavaScript, e.g. `div` in `div { }`.
+ *
+ * @returns The {@link MangleExpression}s.
+ */
+function newElementSelectorExpression(): Iterable<MangleExpression> {
+  return [
+    ...newQuerySelectorExpressions({
+      kind: "element",
+    }),
+  ];
+}
+
+/**
+ * Get {@link MangleExpression}s to match query selectors for IDs in JavaScript,
+ * e.g. `foobar` in `#foobar { }`.
+ *
+ * @returns The {@link MangleExpression}s.
+ */
+function newIdSelectorExpression(): Iterable<MangleExpression> {
+  return [
+    ...newQuerySelectorExpressions({
+      kind: "id",
+      prefix: /#/.source,
+    }),
+    ...newSelectorAsStandaloneStringExpressions(),
+  ];
+}
+
+/**
+ * Get the set of {@link MangleExpression}s to match query selectors in
+ * JavaScript. This will match:
+ * - Attribute selectors (e.g. `foobar` in `querySelector("[foobar]")`).
+ * - Class selectors (e.g. `foobar` in `querySelector(".foobar")`).
+ * - Element selectors (e.g. `div` in `querySelector("div")`).
+ * - ID selectors (e.g. `foobar` in `querySelector("#foobar")`).
+ * - Standalone strings (e.g. `foobar` in `getElementById("foobar");`).
+ *
+ * If no `kind` is specified, this will fall back to the behaviour of v0.1.28 of
+ * the plugin.
+ *
+ * @param options The {@link QuerySelectorOptions}.
+ * @returns A set of {@link MangleExpression}s.
+ */
+function querySelectorExpressionFactory(
+  options: QuerySelectorOptions,
+): Iterable<MangleExpression> {
+  switch (options.kind) {
+  case "attribute":
+    return newAttributeSelectorExpression();
+  case "class":
+    return newClassSelectorExpression();
+  case "element":
+    return newElementSelectorExpression();
+  case "id":
+    return newIdSelectorExpression();
+  case undefined:
+    return fallback(options);
+  }
 }
 
 export default querySelectorExpressionFactory;